projects/capstone/starter/requirements.txt
alembic==1.4.2
aniso8601==6.0.0
appdirs==1.4.4
appnope==0.1.0
awscli==1.18.97
Babel==2.8.0
backcall==0.2.0
botocore==1.17.20
certifi==2020.6.20
Click==7.0
colorama==0.4.3
decorator==4.4.2
distlib==0.3.1
docutils==0.15.2
ecdsa==0.15
filelock==3.0.12
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.18.2
gunicorn==20.0.4
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.10.1
jmespath==0.10.0
jose==1.0.0
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
npm==0.1.1
optional-django==0.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pycryptodome==3.3.1
Pygments==2.6.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.4
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.27
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
